% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_activity.R
\name{activity_choice_message}
\alias{activity_choice_message}
\title{Message for what activities they can choose}
\usage{
activity_choice_message()
}
\description{
Message for what activities they can choose
}
