% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{activity_path}
\alias{activity_path}
\title{Path to activities}
\format{
An object of class \code{character} of length 1.
}
\usage{
activity_path
}
\description{
Path to activities
}
\keyword{datasets}
