% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_homework.R
\name{check_homework}
\alias{check_homework}
\title{Let them know if they have bugs - NOTHING FANCY!!}
\usage{
check_homework(number = NULL)
}
\arguments{
\item{number}{The number homework we're on}
}
\value{
invisible number
}
\description{
Let them know if they have bugs - NOTHING FANCY!!
}
