% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%+\%}
\alias{\%+\%}
\title{Crayon operator}
\usage{
lhs \%+\% rhs
}
\arguments{
\item{lhs}{crayonable string}

\item{rhs}{crayonable string}
}
\value{
The strings....
}
\description{
Crayon operator
}
\keyword{internal}
