% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_activity.R
\name{launch_activity}
\alias{launch_activity}
\title{Run activity based on given argument}
\usage{
launch_activity(choice = NULL)
}
\arguments{
\item{choice}{What to to run}
}
\value{
invisible choice or FALSE if failed
}
\description{
Run activity based on given argument
}
