% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_exercises.R
\name{launch_exercises}
\alias{launch_exercises}
\title{Run learnr module or shiny app based on given argument}
\usage{
launch_exercises(choice = NULL)
}
\arguments{
\item{choice}{What to to run}
}
\value{
invisible choice or FALSE if failed
}
\description{
Run learnr module or shiny app based on given argument
}
