% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_homework.R
\name{launch_homework}
\alias{launch_homework}
\title{Create and populate a homework directory}
\usage{
launch_homework(number = NULL)
}
\arguments{
\item{number}{The number homework we're on}
}
\value{
invisible number
}
\description{
Create and populate a homework directory
}
